moved model_db
